requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
Pillow>=10.1.0  # Updated to be compatible with scikit-image
transparent-background==1.3.3
torch>=2.2.0
numpy>=1.24.3
tqdm>=4.66.1
